extract pythia by dataset

+++ distribution_validating.py
@@ -38,7 +38,7 @@ model = GPTNeoXForCausalLM.from_pretrained(
   "EleutherAI/pythia-70m-deduped",
   revision="step143000",
   cache_dir="./pythia-160m-deduped/step143000",
-)
+).half()
 model = model.to_bettertransformer()
 model = model.cuda()
 tokenizer = AutoTokenizer.from_pretrained(
